NEURAL MIND NET USA OFFICIAL WRITTEN DISCREET OR PUBLIC GOVERNMENT ACKNOWLEDGMENT
AUTHOR: 
Dr JOSE ARREOLA THE FOUNDING FATHER OF THE UNITED STATES OF AMERICA (JEW DESCENT) PRESIDENTIAL (neural net agreement on gold)
THEME:
NEURAL MIND NET USA OFFICIAL WRITTEN DISCREET OR PUBLIC GOVERNMENT ACKNOWLEDGMENT 
CATEGORY:
On the effect of information on human comportment and expected conduct

Date of Issuance, 7th of November 2024, Updated
Location of Issuance, On his temporary residence in San Martin Hidalgo, Jalisco, MEXICO

A message from The Founding Father of The United States of America Dr Jose Arreola, also The President of The United States of America under the US Neural Mind Net:

It has come to my attention the consideration of the possibility of Americans to have the NEURAL MIND NET that we use to communicate without the need to talk using our mouths instead only using our minds, having neural gluey air as the medium to transmit information, emotion and imagery, TERMINATED. For that to happen it would first be necessary an OFFICIAL WRITTEN DISCREET OR PUBLIC GOVERNMENT ACKNOWLEDGMENT of its existence and of its implementation on the human brain (mind) primarily and on the entire physique for multiple purposes by The United States of America. 

	An OFFICIAL WRITTEN DISCREET GOVERNMENT ACKNOWLEDGMENT, based on US Congress Consensus and Agreement and the approval of The Executive, would be the first step with a CLASSIFIED DOCUMENT indicating the NEURAL MIND NET IS REAL and that is has been used for centuries in The United States of America. The Government of The United States of America would inform of the realization and pen signing of this CLASSIFIED DOCUMENT.

	An OFFICIAL WRITTEN PUBLIC GOVERNMENT ACKNOWLEDGMENT based on US Congress Consensus and Agreement and the approval of The Executive, would be the second step with an OFFICIAL PUBLIC GOVERNMENT ACKNOWLEDGMENT made by The current in Office President and Vice President of The The United States of America, and The Founding Father of The United States of America Dr Jose Arreola, COVERED BY THE USA MEDIA.

	It is not possible to seriously consider the possibility of permanently ending the neural mind net unless any of the above are carried out. 

	In the past the ‘American Civils’ did not imagine the possibility of permanently ending the neural mind net because they did not have the telecommunication tools we now have such as Internet, Email, Social Media, Cellphone, and Safe Postage. Besides, in the past there was not advanced fire arms and mobile and static artillery that would make the neural mind net more hazardous than ever to human life. Now we have the tools and we can dream of ending permanently the neural mind net from Earth, not from other Plantes.

	For a nation with Imagination under God.


Jose Arreola
Founding Father of The United States of America
President of The United States of America under the neural net 
Member of Congress of The United States of America under the Neural Net, House of Representatives, State of California
Doctor of Philosophy, Dr  
Freedom Honor, US Government, Official 
Medal of Honor, US Army
Cross of Honor For Military Service Abroad, US Army
Leader of the Movement of Neural Independence of the United States of America from Foreign Influence
Leader of the Movement of National Restoration of The United States of America
Official Dismantler of the neural mind net of the US and the World  
"The quest of our nation is to protect freedom and other rights and liberties that have led us to be what we are, a beacon of hope to the world in times of despair, of discord and of global predatory self-deserving, and to care for the common good when relative interest of supremacy claim go egocentric" Founding Father of the United States of America Dr Jose Arreola
;
The Category of  ‘PRESIDENTIAL’ was added to this document on the date specified above.