Create Mobile Testing Cars for HIV and AIDS Detection
AUTHOR: 
Dr JOSE ARREOLA THE FOUNDING FATHER OF THE UNITED STATES OF AMERICA (JEW DESCENT) PRESIDENTIAL (neural net agreement on gold)
THEME:
Mobile Testing Cars for HIV and AIDS Detection
CATEGORY:
On the effect of information on human comportment and expected conduct

Date of Issuance, 24th of November 2024, Updated
Location of Issuance, On his temporary residence in Los Angeles, California, USA

A message from The Founding Father of The United States of America Dr Jose Arreola, also The President of The United States of America under the US Neural Mind Net:

The purpose of my email is to strongly suggest the starting of a Health Campaign to identify, combat, manage and mitigate the risk of transmitting viral contagious diseases such as HIV and its development AIDS (Acquired Immunodeficiency Syndrome, commonly showing on the skin as “Sarcoma de Kaposi”) that have no cure and have the potential of being spread to all age categories under neural net Societies and their sex culture. This under the context of the most recent developed sex habits of American Citizens in circumstances of local and foreign neural nets and their political and non-political representatives' influence. A practical approach to identifying, combating, managing and mitigating the risk of transmitting HIV and its development AIDS is by assigning around 2600 (two thousand six hundred) “Mobile Testing Cars for HIV and AIDS Detection” in all States of the Country. The epidemic of drugs our country is facing and the neglecting attitude of drug users to having sex carelessly without using condom or other type of protection against viral diseases is a factor that is largely contributing to the steady and rapid spread of this non benign illnesses. Once diagnosis of HIV and and its development AIDS has been conducted, people found positive should be incorporated to the records and data base of “Internet Base Health Platform Application” proposed by Jose Arreola our 8th Founding Father of the United States of America that will potentially be developed by a renown university's  School of Medicine, for instance and will serve the eradication and management of venereal diseases including HIV and its development AIDS, for safer intercourse sex relationships motivation, and for the healthy propagation of the human species for future generations.

For a nation with Imagination under God.


Jose Arreola
Founding Father of The United States of America
President of The United States of America under the neural net 
Member of Congress of The United States of America under the Neural Net, House of Representatives, State of California
Doctor of Philosophy, Dr  
Freedom Honor, US Government, Official 
Medal of Honor, US Army
Cross of Honor For Military Service Abroad, US Army
Leader of the Movement of Neural Independence of the United States of America from Foreign Influence
Leader of the Movement of National Restoration of The United States of America
Official Dismantler of the neural mind net of the US and the World  
"The quest of our nation is to protect freedom and other rights and liberties that have led us to be what we are, a beacon of hope to the world in times of despair, of discord and of global predatory self-deserving, and to care for the common good when relative interest of supremacy claim go egocentric" Founding Father of the United States of America Dr Jose Arreola