Create Pro Correction of Law Enforcement Practices and Records
AUTHOR: 
Dr JOSE ARREOLA THE FOUNDING FATHER OF THE UNITED STATES OF AMERICA (JEW DESCENT) PRESIDENTIAL (neural net agreement on gold)
THEME:
Pro Correction of Law Enforcement Practices and Records
CATEGORY:
Matching Video Footage Surveillance with Missing Police Department's Arrest File Records using Facial Recognition

A message from The Founding Father of The United States of America Dr Jose Arreola, also The President of The United States of America under the US Neural Mind Net,

It has come to my attention the consideration of improving Law Enforcement Practices in oder to correct the mishandling of numerous cases of people who went to prison at least once in the USA’s prison setup in the past 4 years (2020-2025), according to information extracted from the collective memory using the US neural mind net, and were released from prison due to the influence of neural mind nets (communication and influence on and through the mind) local and from abroad, without leaving a record of their temporary imprisonment, hourly in most instances, in the computerized system of any Law Enforcement Departments. Missing file records of people who went to prison in the past in the USA’s prison setup would be registered in the computerized system of any Law Enforcement Departments to match them with the number of times they were in prison and the reasons why. The missing file records would contain Prison Cell time and Sherif’s Patrol’s time when being taken to jail and would be obtained from the recorded camera video footage surveillance of Prisons’ Cell time and Sherif’s Patrols. Redone file records could served for multiple purposes such as the consideration for re-imprisonment and would in turn increase safety on the streets of America while removing for a longer period of time, or permanently, the number of people who went to prison in the past at least once, if no record of their temporary imprisonment was left, and if they happen to go to prison again due to incorrigible unlawful misconduct. 

For a nation with Imagination under God.


Jose Arreola
Founding Father of The United States of America
President of The United States of America under the neural net 
Member of Congress of The United States of America under the Neural Net, House of Representatives, State of California
Doctor of Philosophy, Dr  
Freedom Honor, US Government, Official 
Medal of Honor, US Army
Cross of Honor For Military Service Abroad, US Army
Leader of the Movement of Neural Independence of the United States of America from Foreign Influence
Leader of the Movement of National Restoration of The United States of America
Official Dismantler of the neural mind net of the US and the World  
"The quest of our nation is to protect freedom and other rights and liberties that have led us to be what we are, a beacon of hope to the world in times of despair, of discord and of global predatory self-deserving, and to care for the common good when relative interest of supremacy claim go egocentric" Founding Father of the United States of America Dr Jose Arreola
;
The Category of  ‘PRESIDENTIAL’ was added to this document on the date specified above.