Create Internet Base Health Platform Application
AUTHOR: 
Dr JOSE ARREOLA THE FOUNDING FATHER OF THE UNITED STATES OF AMERICA (JEW DESCENT) PRESIDENTIAL (neural net agreement on gold)
THEME:
Internet Base Health Platform Application
CATEGORY:
On the effect of information on human comportment and expected conduct

Date of Issuance, 24th of November 2024, Updated
Location of Issuance, On his temporary residence in Los Angeles, California, USA

A message from The Founding Father of The United States of America Dr Jose Arreola, also The President of The United States of America under the US Neural Mind Net:

I am contacting you to convene to participate in a project to set up in the USA an Internet Base Health Platform Application where all Americans would have a health profile indicating the type and number of venereal diseases, including AIDS  they have, with the purpose of ensuring safer intercourse sex relationships, the gradual eradication and management of venereal diseases such as AIDS, undergoing lower costs for Federal and State Health Care Programs, and to ensure in the long term, for future generations, a healthy propagation of the human species. By login into the Internet Base Health Platform Application in your Cellphone, after having downloaded it, through granted or without requiring permission, a person finds out if someone has acquired a certain venereal disease such as AIDS or others, and the number of them. Then, a person decides after wining and dining whether to have or not have intercourse sex relationships with someone, having had access to the information provided by The Internet Base Health Platform Application, leading to better outcomes to overall Health. The Internet Base Health Platform Application would be financed by each person at download, and the setting up and management of it could be carried out by a private Enterprise.

For a nation with Imagination under God.

Jose Arreola
Founding Father of The United States of America
President of The United States of America under the neural net 
Member of Congress of The United States of America under the Neural Net, House of Representatives, State of California
Doctor of Philosophy, Dr  
Freedom Honor, US Government, Official 
Medal of Honor, US Army
Cross of Honor For Military Service Abroad, US Army
Leader of the Movement of Neural Independence of the United States of America from Foreign Influence
Leader of the Movement of National Restoration of The United States of America
Official Dismantler of the neural mind net of the US and the World  
"The quest of our nation is to protect freedom and other rights and liberties that have led us to be what we are, a beacon of hope to the world in times of despair, of discord and of global predatory self-deserving, and to care for the common good when relative interest of supremacy claim go egocentric" Founding Father of the United States of America Dr Jose Arreola