Create The USA Culture Design and Enhancement Project
AUTHOR: 
Dr JOSE ARREOLA THE FOUNDING FATHER OF THE UNITED STATES OF AMERICA (JEW DESCENT) PRESIDENTIAL (neural net agreement on gold)
THEME:
The USA Culture Design and Enhancement Project
CATEGORY:
On the effect of information on human comportment and expected conduct

Date of Issuance, 24th of November 2024, Updated
Location of Issuance, On his temporary residence in Los Angeles, California, USA

A message from The Founding Father of The United States of America Dr Jose Arreola, also The President of The United States of America under the US Neural Mind Net:

I am contacting you to invite you to participate in what we conceive as ‘The USA Culture Design and Enhancement Project’ given the events that have occurred in the last1 year and a half in the USA in terms of the incorporation of novel daily habits from neural net-context-based sources and environments, and new adventures that came from afar. All of the above with the purpose of reasserting and enhancing our American identity and culture and its potential for guaranteeing the freedoms we all ambition, and entailed pleasures, to enjoy within the frame of the law in a Capitalist Model country that stresses had work and productivity. A culture must not stop developing and enriching in a world of continuous change, of increasing existential needs for meaning of form and matter (psyche and physique). The development of our culture should not be left to the random motions and stimuli of the collective consciousness and passions of the group in our Cities and States of the Nation underlying criteria not aligned to the Principles of Milton Freeman of the Free Markets that yield the most optimal and fair outcomes. Instead our culture must be carefully reassessed, bounded in some directions, rid off in other directions, and redesigned and enriched with new elements. The setting of new infrastructures and new IT technologies and all type of innovations such as Public Parks, automations and robotics emerging for more efficient knowledge harnessing (i.e. Self-Education Prisons suggested by Dr. Jose Arreola Hernandez, The Founding Father of the United States of America) and acquisitions should be incorporated as part of the Culture Design Project of the USA. City Planners, Designers and Marketers could work together on this endeavor.

For a nation with Imagination under God.


Jose Arreola
Founding Father of The United States of America
President of The United States of America under the neural net 
Member of Congress of The United States of America under the Neural Net, House of Representatives, State of California
Doctor of Philosophy, Dr  
Freedom Honor, US Government, Official 
Medal of Honor, US Army
Cross of Honor For Military Service Abroad, US Army
Leader of the Movement of Neural Independence of the United States of America from Foreign Influence
Leader of the Movement of National Restoration of The United States of America
Official Dismantler of the neural mind net of the US and the World  
"The quest of our nation is to protect freedom and other rights and liberties that have led us to be what we are, a beacon of hope to the world in times of despair, of discord and of global predatory self-deserving, and to care for the common good when relative interest of supremacy claim go egocentric" Founding Father of the United States of America Dr Jose Arreola